+++ pkg/mem/set.arm64.go
@@ -6,16 +6,16 @@ package mem
 // `n` iterations, with the value `v`.
 //
 //go:noescape
-func setU8(a *byte, n uint64, v uint8)
+func setU8(a *byte, n int, v uint8)
 
 // setU32 sets every four bytes, starting at `a` and going on for
 // `n` iterations, with the value `v`.
 //
 //go:noescape
-func setU32(a *byte, n uint64, v uint32)
+func setU32(a *byte, n int, v uint32)
 
 // setU64 sets every eight bytes, starting at `a` and going on for
 // `n` iterations, with the value `v`.
 //
 //go:noescape
-func setU64(a *byte, n uint64, v uint64)
+func setU64(a *byte, n int, v uint64)
